[llvm] [refactor] LLVMProgramImpl code clean up: part-4
